+++ src/lib.rs
@@ -61,6 +61,7 @@ pub use crate::algorithms::io::edgelist::read_edgelist;
 pub use crate::algorithms::operators::complementer::complementer;
 pub use crate::algorithms::operators::disjoint_union::{disjoint_union, disjoint_union_many};
 pub use crate::algorithms::operators::simplify::simplify;
+pub use crate::algorithms::operators::union::union;
 pub use crate::algorithms::paths::dijkstra::dijkstra_distances;
 pub use crate::algorithms::paths::distances::distances;
 pub use crate::algorithms::paths::eulerian::{EulerianClassification, is_eulerian};